% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directions.R
\name{directions}
\alias{directions}
\title{Finding the directions}
\usage{
directions(origin, destination, avoid = NULL, travel_mode = "driving")
}
\arguments{
\item{origin}{a character string for the name of the origin}

\item{destination}{a character string for the name of the destination}

\item{avoid}{an optional parameter using to indicate that the calculated route should avoid this typical features.
Followings are the list of features:
\itemize{
 \item tolls
 \item highways
 \item ferries
}}

\item{travel_mode}{an optional parameter using to indicate the mode of travel.
\itemize{
 \item \strong{driving} (default): indicates the driving direction on the road.
 \item \strong{walking}: indicates walking direction on the pedestrian.
 \item \strong{bicycling}: indicates the bicycling direction on the bicycle path.
}}
}
\value{
The map showing the origin and destination places along with
the several red dots which indicate the intersection or crossing between the route.
}
\description{
A function to find the route between two countries/cities.
}
\examples{
directions("Vienna", "Graz")
directions("Bangkok", "Pattaya", avoid="tolls")
directions("Manchester", "Leeds", travel_mode="walking")
}

